% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gmiShape}
\alias{gmiShape}
\title{Kształty gmin}
\format{
Ramka danych stworzona funkcją \code{read_sf}
\describe{
  \item{teryt}{kod teryt gminy w formacie czterocyfrowym}
  \item{nazwa}{nazwy gmin}
  \item{geometry}{kształty gmin potrzebne do narysowania mapy}
 }
}
\usage{
gmiShape
}
\description{
Ramka danych zawierająca kształty gmin wymagane do rysowania
map oraz informacje potrzebne do zidentyfikowania gmin takie, jak kod teryt,
nazwy gmin itd.
}
\keyword{datasets}
