% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{ibe_theme}
\alias{ibe_theme}
\title{Motyw wykorzystywany w publikacjach w projekcie Losy 2 (IBE)}
\usage{
ibe_theme()
}
\description{
Minimalistyczny motyw używany w raportach automatycznych w
projekcie Losy 2 w ramach IBE.
}
