% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{makroregShape}
\alias{makroregShape}
\title{Kształty makroregionów (NUTS1)}
\format{
Ramka danych w formacie \code{data.frame}
\describe{
  \item{nazwa}{nazwa makroregionu}
  \item{kod}{kod makroregionu składający się z liter "PL" i jednej cyfry}
  \item{geometry}{kształty makroregionów potrzebne do narysowania mapy}
 }
}
\usage{
makroregShape
}
\description{
Ramka danych zawierająca kształty makroregionów według
klasyfikacji NUTS1. Można ich używać do nakładania nowej warstwy na wykres
(np. granic makroregionów w postaci konturu lub wypełnienia kolorem z
parametrem przezroczystości \code{alpha})
}
\keyword{datasets}
