% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapa_cont.R
\name{mapa_gmina_cont}
\alias{mapa_gmina_cont}
\title{Mapa z terytorialnym rozkładem \strong{zmiennej ciągłej} na poziomie
gmin}
\usage{
mapa_gmina_cont(x, mapping, teryt_var)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami zagregowanymi na poziomie województw}

\item{mapping}{nazwa zmiennej, której rozkład ma być przedstawiony na
wykresie}

\item{teryt_var}{nazwa zmiennej zawierająca numer teryt województwa w formie
dwuocyfrowej ze zbioru danych przekazanego do argumentu \code{x}}
}
\description{
Funkcja łączy ramkę danych zawierającą kształty (shape'y)
gmin z dowolną ramką danych zawierającą zmienną \code{teryt_gmi} - właśnie po
tej zmiennej, a następnie zwraca wykres w formie mapy z rozkładem zmiennej
ciągłej. Obszary gmin kolorowane są gradientem w zależności od wartości
zmiennej.
}
