% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{podregShape}
\alias{podregShape}
\title{Kształty podregionów (NUTS3)}
\format{
Ramka danych w formacie \code{data.frame}
\describe{
  \item{nazwa}{nazwa podregionu}
  \item{kod}{kod podregionu składający się z liter "PL" i trzech cyfr}
  \item{geometry}{kształty podregionów potrzebne do narysowania mapy}
 }
}
\usage{
podregShape
}
\description{
Ramka danych zawierająca kształty podregionów według
klasyfikacji NUTS3. Można ich używać do nakładania nowej warstwy na wykres
(np. granic podregionów w postaci konturu lub wypełnienia kolorem z
parametrem przezroczystości \code{alpha})
}
\keyword{datasets}
