% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{powiatyShape}
\alias{powiatyShape}
\title{Kształty powiatów}
\format{
Ramka danych stworzona funkcją \code{read_sf}
\describe{
  \item{teryt}{kod teryt powiatu w formacie czterocyfrowym}
  \item{nazwa}{nazwy powiatów}
  \item{geometry}{kształty powiatów potrzebne do narysowania mapy}
 }
}
\usage{
powiatyShape
}
\description{
Ramka danych zawierająca kształty powiatów wymagane do rysowania
map oraz informacje potrzebne do zidentyfikowania powiatów takie, jak kod
teryt, nazwy powiatów itd.
}
\keyword{datasets}
