% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teryt_recode.R
\name{teryt_recode_pow}
\alias{teryt_recode_pow}
\title{zmiana formatu terytu powiatu}
\usage{
teryt_recode_pow(x)
}
\arguments{
\item{x}{kolumna zawierajaca numer teryt - trzeci argument funkcji
\code{mapa_powiat_cont}}
}
\description{
funkcja konwertuje różne formaty terytu dla powiatu na
kompatybilny z funkcją łączącą zbiory i rysującą podstawowy wykres
}
