% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teryt_recode.R
\name{teryt_recode_woj}
\alias{teryt_recode_woj}
\title{zmiana formatu terytu województwa}
\usage{
teryt_recode_woj(x)
}
\arguments{
\item{x}{kolumna zawierajaca numer teryt - trzeci argument funkcji
\code{mapa_woj_cont}}
}
\description{
funkcja konwertuje różne formaty terytu dla województwa na
kompatybilny z funkcją łączącą zbiory i rysującą podstawowy wykres
}
