% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wojShape}
\alias{wojShape}
\title{Kształty województw}
\format{Ramka danych stworzona funkcją \code{read_sf}
\describe{
  \item{jpt_kod_je}{kod teryt województwa w formacie dwucyfrowym}
  \item{jpt_nazwa_}{nazwy województw}
  \item{geometry}{kształty województw potrzebne do narysowania mapy}
 }}
\usage{
wojShape
}
\description{
Ramka danych zawierająca kształty województw wymagane do
rysowania map oraz informacje potrzebne do zidentyfikowania województw takie,
jak kod teryt, nazwy województw itd.
}
\keyword{datasets}
