% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wschZachShape}
\alias{wschZachShape}
\title{Kształty podziału na Polskę wschodnią i zachodnią}
\format{
Ramka danych w formacie \code{data.frame}
\describe{
  \item{geometry}{kształty obszarów potrzebne do narysowania mapy}
  \item{nazwa}{nazwa obszaru: "wschod" lub "zachod"}
 }
}
\usage{
wschZachShape
}
\description{
Ramka danych zawierająca kształty podziału na Polskę wschodnią i
zachodnią - Polska wschodnia definiowana jest według definicji w programie
Polska Wschodnia 2014-2020 (PO PW), czyli w jej skład wchodzą województwa:
lubelskie, podlaskie, podkarpackie, świętokrzyskie i warmińsko-mazurskie,
natomiast Polska zachodnia to pozostałe województwa. KonturóW można używać do
nakładania nowej warstwy na wykres (np. granic podziału w postaci konturu lub
wypełnienia kolorem z parametrem przezroczystości \code{alpha})
}
\keyword{datasets}
