% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zaboryShape}
\alias{zaboryShape}
\title{Kształty zaborów (bez kategorii ziemie odzyskane)}
\format{
Ramka danych w formacie \code{data.frame}
\describe{
  \item{geometry}{kształty zaborów potrzebne do narysowania mapy}
  \item{nazwa}{nazwa zaboru: "austriacki", "pruski" lub "rosyjski"}
 }
}
\usage{
zaboryShape
}
\description{
Ramka danych zawierająca kształty zaborów. KonturóW można używać
do nakładania nowej warstwy na wykres (np. granic zaborów w postaci konturu
lub wypełnienia kolorem z parametrem przezroczystości \code{alpha})
}
\keyword{datasets}
