% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zaboryZOShape}
\alias{zaboryZOShape}
\title{Kształty zaborów (zawierające kategorię ziem odzyskanych)}
\format{
Ramka danych w formacie \code{data.frame}
\describe{
  \item{geometry}{kształty zaborów potrzebne do narysowania mapy}
  \item{nazwa}{nazwa zaboru: "austriacki", "pruski", "rosyjski" lub ziemie
  odzyskane}
 }
}
\usage{
zaboryZOShape
}
\description{
Ramka danych zawierająca kształty zaborów z dodatkową kateogią
ziem odzyskanych. KonturóW można używać do nakładania nowej warstwyn a wykres
(np. granic zaborów w postaci konturu lub wypełnienia kolorem z parametrem
przezroczystości \code{alpha})
}
\keyword{datasets}
