% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapa_cont.R
\name{mapa_woj_cont}
\alias{mapa_woj_cont}
\title{Mapa z terytorialnym rozkładem \strong{zmiennej ciągłej} na poziomie
województw}
\usage{
mapa_woj_cont(x, mapping, var)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami zagregowanymi na poziomie województw}

\item{mapping}{nazwa zmiennej wpisana jako string, która ma być przedstawiona
na wykresie}

\item{var}{nazwa zmiennej zawierająca numer teryt województwa w formie
dwuocyfrowej ze zbioru danych przekazanego do argumentu \code{x}}
}
\description{
Funkcja łączy ramkę danych zawierającą kształty (shape'y)
województw z dowolną ramką danych zawierającą zmienną \code{teryt_woj} -
 właśnie po tej zmiennej, a następnie zwraca wykres w formie mapy z rozkładem
zmiennej ciągłej. Obszary województw kolorowane są gradientem w zależności od
wartości zmiennej.
}
